Lots of changes.
Lcd display homeScreen now displays the temp, and the status of the
  defroster and evaporator.
Buttons and a menu where added so the user interacts with the config.

The menu items can be moved up and down to be able to interact with all
of them and there's a selector arrow that shows the currently selected
one, but they are still just placeholders that do nothing.

Next step will be to make them functional.